Cache layout component types dependent on props in their components rather than recreating them on every render
Recreating the component type causes the React to think that a new
component is appearing on every render, forcing it to replace the old
component with the new one. While this is likely pretty bad for
performance, since there’s a lot of DOM nodes getting put in or taken
out, it also removes any focus on the previous element (and hence
making input elements totally unusable).